requirements.txt
python-telegram-bot
openai
anthropic
python-dotenv
tavily-python
